Replacement filter (#2)
Implements a `replacementFilterRegex` that allows you to limit which files are scanned for old hashes for replacement. This is a performance optmisation that minimises the work the plugin needs to do. Enabling this option enables a build failing safety net that will scan all the assets once at the end for any "old" hashes and fail if they're found (this means you probably missed a file).